requirements.txt
fastapi[all]
urllib3
starlette
paramiko
pandas
requests
passlib
jwt
urllib3
PyJWT
toml
ruamel.yaml
wordcloud
matplotlib
numpy
Pillow
